pyproject.toml
[project]
name = "e-commerce-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioredis>=2.0.1",
    "aiosqlite>=0.22.1",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt==4.1.2",
    "disutils>=1.4.32.post2",
    "fastapi>=0.127.0",
    "greenlet>=3.3.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "redis[asyncio]>=7.1.0",
    "setuptools>=80.9.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]
